% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.cite_author}
\alias{.cite_author}
\title{Format author names for short citation from epidemiological parameter
database}
\usage{
.cite_author(x)
}
\arguments{
\item{x}{A single \code{list} entry from the list of epidemiological parameter
database.}
}
\value{
A \code{character} string with the short citation.
}
\description{
Format author names for short citation from epidemiological parameter
database
}
\keyword{internal}
