% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{.is_pkg_installed}
\alias{.is_pkg_installed}
\title{Check whether a package is installed}
\usage{
.is_pkg_installed(package)
}
\arguments{
\item{package}{string naming the package/name space to load.}
}
\value{
Invisibly returns a boolean \code{logical}.
}
\description{
Check whether a package is installed
}
\details{
This functions allows testing for when a suggested package
dependency is not installed to check if conditional code can be run.
}
\keyword{internal}
