% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{.read_db}
\alias{.read_db}
\title{Reads in parameter library from JSON}
\usage{
.read_db()
}
\value{
\code{list}.
}
\description{
Read epidemiological parameter database from JSON using
\code{\link[jsonlite:read_json]{jsonlite::read_json()}} and returns the list with a \verb{<epiparameterDB>}
class attribute to validate it has been read using this function.
}
\keyword{internal}
