% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{epiparameterDB-package}
\alias{epiparameterDB-package}
\title{epiparameterDB: Database of epidemiological parameters}
\description{
Your package description. It must end with a period (".") and include relevant bibliographical references if applicable, using the following format: Author et al. (2023) \doi{10.5281/zenodo.6619350}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/epiverse-trace/epiparameterDB/}
  \item \url{https://epiverse-trace.github.io/epiparameterDB/}
  \item Report bugs at \url{https://github.com/epiverse-trace/epiparameterDB/issues}
}

}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adam Kucharski \email{adam.kucharski@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-8814-9421}{ORCID})
  \item Carmen Tamayo \email{carmen.tamayo-cuartero@lshtm.ac.uk} (\href{https://orcid.org/0000-0003-4184-2864}{ORCID})
}

}
\keyword{internal}
