% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{epiparameterDB}
\alias{epiparameterDB}
\title{A list of epidemiological parameters.}
\format{
\subsection{\code{epiparameterDB}}{

A \code{list} of epidemiological parameters read from \code{extdata/parameters.json}
with 133 database entries. The object
is tagged with the class attribute \verb{<epiparameterDB>} to validate the
parameter library has been loaded using the \code{{epiparameterDB}} package.
}
}
\usage{
epiparameterDB
}
\description{
A list of epidemiological parameters.
}
\keyword{datasets}
